Update pages to reflect required auth on the client-side
