+++ examples/auth0/README.md
@@ -69,7 +69,7 @@ This sample tries to cover a few topics:
 
 - Signing in
 - Signing out
-- Loading the user on the server side and adding it as part of SSR (`/pages/profile.js`)
+- Loading the user on the server side and adding it as part of SSR (`/pages/advanced/srr-profile.js`)
 - Loading the user on the client side and using fast/cached SSR pages (`/pages/index.js`)
 - API Routes which can load the current user (`/pages/api/me.js`)
-- Using hooks to make the user available throughout the application (`/lib//user.js`)
+- Using hooks to make the user available throughout the application (`/lib/user.js`)
